Rename Brick.hpp -> Rectilinear.hpp
Only do the rename so source control keeps track of the file changes.
