+++ src/Domain/Creators/CMakeLists.txt
@@ -11,7 +11,6 @@ spectre_target_sources(
   AlignedLattice.cpp
   BinaryCompactObject.cpp
   BlockGroups.cpp
-  Brick.cpp
   Cylinder.cpp
   CylindricalBinaryCompactObject.cpp
   Disk.cpp
@@ -19,6 +18,7 @@ spectre_target_sources(
   FrustalCloak.cpp
   Interval.cpp
   Rectangle.cpp
+  Rectilinear.cpp
   RegisterDerivedWithCharm.cpp
   RotatedBricks.cpp
   RotatedIntervals.cpp
@@ -33,7 +33,6 @@ spectre_target_headers(
   AlignedLattice.hpp
   BinaryCompactObject.hpp
   BlockGroups.hpp
-  Brick.hpp
   Cylinder.hpp
   CylindricalBinaryCompactObject.hpp
   Disk.hpp
@@ -48,6 +47,7 @@ spectre_target_headers(
   Interval.hpp
   OptionTags.hpp
   Rectangle.hpp
+  Rectilinear.hpp
   RegisterDerivedWithCharm.hpp
   RotatedBricks.hpp
   RotatedIntervals.hpp
