+++ tests/Unit/IO/Observers/Test_VolumeObserver.cpp
@@ -16,7 +16,7 @@
 
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Index.hpp"
-#include "Domain/Creators/Brick.hpp"
+#include "Domain/Creators/Rectilinear.hpp"
 #include "Domain/Creators/RegisterDerivedWithCharm.hpp"
 #include "Domain/Creators/Tags/Domain.hpp"
 #include "Domain/Creators/Tags/FunctionsOfTime.hpp"
