+++ tests/Unit/ParallelAlgorithms/ApparentHorizonFinder/Test_ComputeHorizonVolumeQuantities.cpp
@@ -8,7 +8,7 @@
 #include <memory>
 
 #include "Domain/Block.hpp"
-#include "Domain/Creators/Brick.hpp"
+#include "Domain/Creators/Rectilinear.hpp"
 #include "Domain/Domain.hpp"
 #include "Domain/ElementMap.hpp"
 #include "Domain/Structure/ElementId.hpp"
